Restore recursive component tests.
Fix PlayerStatus test which really tested something in PlayerInfo.
